@@ -4,5 +4,5 @@ namespace BeyondCode\LaravelMaskedDumper\Contracts;
 
 interface Column
 {
-    public function modifyValue($value);
+    public function modifyValue($value, $rows);
 }
